refactor: reorganize classes inside proper packages
